std related fix for model files

+++ src/problems/LaPalisse.cu
@@ -244,9 +244,9 @@ LaPalisse_imposeBoundaryCondition(
 		if (!VEL_IO(info)) {
 			if (object(info)==1) {
 				// rise slowly over 6 seconds
-				waterdepth = (INLET_WATER_LEVEL - 1.08f - INITIAL_WATER_LEVEL)*fmin(t/RISE_TIME, 1.0f) + INITIAL_WATER_LEVEL;
+				waterdepth = (INLET_WATER_LEVEL - 1.08f - INITIAL_WATER_LEVEL)*fminf(t/RISE_TIME, 1.0f) + INITIAL_WATER_LEVEL;
 			}
-			const float localdepth = fmax(waterdepth - absPos.z, 0.0f);
+			const float localdepth = fmaxf(waterdepth - absPos.z, 0.0f);
 			const float pressure = 9.81e3f*localdepth;
 			eulerVel.w = RHO(pressure, fluid_num(info));
 		}
